Implement service loader for PacketFactoryBinder
Github issue: #207
The PacketFactoryBinder was injected in the classpath to be used by the
PacketFactories. This implied to have the same package name but is not
compatible with the Java 9 modularity. I kept the same behavior but
using ServiceLoader instead.
