+++ pcap4j-core/src/main/java/org/pcap4j/packet/factory/PacketFactoryBinder.java
@@ -10,21 +10,19 @@ package org.pcap4j.packet.factory;
 import org.pcap4j.packet.namednumber.NamedNumber;
 
 /**
- * @author Kaito Yamada
- * @since pcap4j 0.9.16
+ * <p>Pcap4J modules can provide a factory to build new packets as they are received.<br>
+ * The implementing modules must also provide a {@link PacketFactoryBinderProvider}</p>
+ * 
+ * @author Jordan Dubie
+ * @since pcap4j 1.7.6
  */
-final class PacketFactoryBinder {
-
-  private static final PacketFactoryBinder INSTANCE = new PacketFactoryBinder();
-
-  private PacketFactoryBinder() {}
-
-  public static PacketFactoryBinder getInstance() {
-    return INSTANCE;
-  }
-
-  public <T, N extends NamedNumber<?, ?>> PacketFactory<T, N> getPacketFactory(
-      Class<T> targetClass, Class<N> numberClass) {
-    throw new UnsupportedOperationException("This code is never included in pcap4j-core.jar.");
-  }
+public interface PacketFactoryBinder {
+    /**
+     * Provides a {@link PacketFactory} to build the received packets.
+     * 
+     * @param targetClass targetClass
+     * @param numberClass numberClass
+     * @return the factory
+     */
+    public <T, N extends NamedNumber<?, ?>> PacketFactory<T, N> getPacketFactory(Class<T> targetClass, Class<N> numberClass);
 }
