+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedIpV4OptionFactory.java
@@ -1,44 +1,45 @@
 /*_##########################################################################
   _##
-  _##  Copyright (C) 2016  Pcap4J.org
+  _##  Copyright (C) 2012-2014  Pcap4J.org
   _##
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import org.pcap4j.packet.IllegalRadiotapData;
+
+import org.pcap4j.packet.IllegalIpV4Option;
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.RadiotapPacket.RadiotapData;
-import org.pcap4j.packet.namednumber.RadiotapPresentBitNumber;
+import org.pcap4j.packet.IpV4Packet.IpV4Option;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.namednumber.IpV4OptionType;
 
 /**
  * @author Kaito Yamada
- * @since pcap4j 1.6.5
+ * @since pcap4j 0.9.14
  */
-public final class PropertiesBasedRadiotapDataFactory
-    implements PacketFactory<RadiotapData, RadiotapPresentBitNumber> {
+public final class PropertiesBasedIpV4OptionFactory
+    implements PacketFactory<IpV4Option, IpV4OptionType> {
 
-  private static final PropertiesBasedRadiotapDataFactory INSTANCE =
-      new PropertiesBasedRadiotapDataFactory();
+  private static final PropertiesBasedIpV4OptionFactory INSTANCE =
+      new PropertiesBasedIpV4OptionFactory();
 
-  private PropertiesBasedRadiotapDataFactory() {}
+  private PropertiesBasedIpV4OptionFactory() {}
 
-  /** @return the singleton instance of PropertiesBasedRadiotapDataFactory. */
-  public static PropertiesBasedRadiotapDataFactory getInstance() {
+  /** @return the singleton instance of PropertiesBasedIpV4OptionFactory. */
+  public static PropertiesBasedIpV4OptionFactory getInstance() {
     return INSTANCE;
   }
 
   @Override
-  public RadiotapData newInstance(
-      byte[] rawData, int offset, int length, RadiotapPresentBitNumber num) {
-    return newInstance(rawData, offset, length, getTargetClass(num));
+  public IpV4Option newInstance(byte[] rawData, int offset, int length, IpV4OptionType number) {
+    return newInstance(rawData, offset, length, getTargetClass(number));
   }
 
   @Override
-  public RadiotapData newInstance(byte[] rawData, int offset, int length) {
+  public IpV4Option newInstance(byte[] rawData, int offset, int length) {
     return newInstance(rawData, offset, length, getTargetClass());
   }
 
@@ -47,14 +48,14 @@ public final class PropertiesBasedRadiotapDataFactory
    * @param offset offset
    * @param length length
    * @param dataClass dataClass
-   * @return a new RadiotapDataField object.
+   * @return a new IpV4Option object.
    * @throws IllegalStateException if an access to the newInstance method of the dataClass fails.
    * @throws IllegalArgumentException if an exception other than {@link IllegalRawDataException} is
    *     thrown by newInstance method of the dataClass.
    * @throws NullPointerException if any of arguments are null.
    */
-  public RadiotapData newInstance(
-      byte[] rawData, int offset, int length, Class<? extends RadiotapData> dataClass) {
+  public IpV4Option newInstance(
+      byte[] rawData, int offset, int length, Class<? extends IpV4Option> dataClass) {
     if (rawData == null || dataClass == null) {
       StringBuilder sb = new StringBuilder(50);
       sb.append("rawData: ").append(rawData).append(" dataClass: ").append(dataClass);
@@ -63,7 +64,7 @@ public final class PropertiesBasedRadiotapDataFactory
 
     try {
       Method newInstance = dataClass.getMethod("newInstance", byte[].class, int.class, int.class);
-      return (RadiotapData) newInstance.invoke(null, rawData, offset, length);
+      return (IpV4Option) newInstance.invoke(null, rawData, offset, length);
     } catch (SecurityException e) {
       throw new IllegalStateException(e);
     } catch (NoSuchMethodException e) {
@@ -74,22 +75,22 @@ public final class PropertiesBasedRadiotapDataFactory
       throw new IllegalStateException(e);
     } catch (InvocationTargetException e) {
       if (e.getTargetException() instanceof IllegalRawDataException) {
-        return IllegalRadiotapData.newInstance(rawData, offset, length);
+        return IllegalIpV4Option.newInstance(rawData, offset, length);
       }
       throw new IllegalArgumentException(e);
     }
   }
 
   @Override
-  public Class<? extends RadiotapData> getTargetClass(RadiotapPresentBitNumber num) {
-    if (num == null) {
-      throw new NullPointerException("num must not be null");
+  public Class<? extends IpV4Option> getTargetClass(IpV4OptionType number) {
+    if (number == null) {
+      throw new NullPointerException("number: " + number);
     }
-    return PacketFactoryPropertiesLoader.getInstance().getRadiotapDataFieldClass(num);
+    return PacketFactoryPropertiesLoader.getInstance().getIpV4OptionClass(number);
   }
 
   @Override
-  public Class<? extends RadiotapData> getTargetClass() {
-    return PacketFactoryPropertiesLoader.getInstance().getUnknownRadiotapDataFieldClass();
+  public Class<? extends IpV4Option> getTargetClass() {
+    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV4OptionClass();
   }
 }
