+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedIpV6NeighborDiscoveryOptionFactory.java
@@ -5,13 +5,15 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+
 import org.pcap4j.packet.IcmpV6CommonPacket.IpV6NeighborDiscoveryOption;
 import org.pcap4j.packet.IllegalIpV6NeighborDiscoveryOption;
 import org.pcap4j.packet.IllegalRawDataException;
+import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.IpV6NeighborDiscoveryOptionType;
 
 /**
