+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedIpV6OptionFactory.java
@@ -5,39 +5,41 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import org.pcap4j.packet.IllegalIpV6RoutingData;
+
+import org.pcap4j.packet.IllegalIpV6Option;
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.IpV6ExtRoutingPacket.IpV6RoutingData;
-import org.pcap4j.packet.namednumber.IpV6RoutingType;
+import org.pcap4j.packet.IpV6ExtOptionsPacket.IpV6Option;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.namednumber.IpV6OptionType;
 
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.14
  */
-public final class PropertiesBasedIpV6RoutingDataFactory
-    implements PacketFactory<IpV6RoutingData, IpV6RoutingType> {
+public final class PropertiesBasedIpV6OptionFactory
+    implements PacketFactory<IpV6Option, IpV6OptionType> {
 
-  private static final PropertiesBasedIpV6RoutingDataFactory INSTANCE =
-      new PropertiesBasedIpV6RoutingDataFactory();
+  private static final PropertiesBasedIpV6OptionFactory INSTANCE =
+      new PropertiesBasedIpV6OptionFactory();
 
-  private PropertiesBasedIpV6RoutingDataFactory() {}
+  private PropertiesBasedIpV6OptionFactory() {}
 
-  /** @return the singleton instance of PropertiesBasedIpV6RoutingDataFactory. */
-  public static PropertiesBasedIpV6RoutingDataFactory getInstance() {
+  /** @return the singleton instance of PropertiesBasedIpV6OptionFactory. */
+  public static PropertiesBasedIpV6OptionFactory getInstance() {
     return INSTANCE;
   }
 
   @Override
-  public IpV6RoutingData newInstance(byte[] rawData, int offset, int length, IpV6RoutingType type) {
-    return newInstance(rawData, offset, length, getTargetClass(type));
+  public IpV6Option newInstance(byte[] rawData, int offset, int length, IpV6OptionType number) {
+    return newInstance(rawData, offset, length, getTargetClass(number));
   }
 
   @Override
-  public IpV6RoutingData newInstance(byte[] rawData, int offset, int length) {
+  public IpV6Option newInstance(byte[] rawData, int offset, int length) {
     return newInstance(rawData, offset, length, getTargetClass());
   }
 
@@ -46,14 +48,14 @@ public final class PropertiesBasedIpV6RoutingDataFactory
    * @param offset offset
    * @param length length
    * @param dataClass dataClass
-   * @return a new IpV6RoutingData object.
+   * @return a new IpV6Option object.
    * @throws IllegalStateException if an access to the newInstance method of the dataClass fails.
    * @throws IllegalArgumentException if an exception other than {@link IllegalRawDataException} is
    *     thrown by newInstance method of the dataClass.
    * @throws NullPointerException if any of arguments are null.
    */
-  public IpV6RoutingData newInstance(
-      byte[] rawData, int offset, int length, Class<? extends IpV6RoutingData> dataClass) {
+  public IpV6Option newInstance(
+      byte[] rawData, int offset, int length, Class<? extends IpV6Option> dataClass) {
     if (rawData == null || dataClass == null) {
       StringBuilder sb = new StringBuilder(50);
       sb.append("rawData: ").append(rawData).append(" dataClass: ").append(dataClass);
@@ -62,7 +64,7 @@ public final class PropertiesBasedIpV6RoutingDataFactory
 
     try {
       Method newInstance = dataClass.getMethod("newInstance", byte[].class, int.class, int.class);
-      return (IpV6RoutingData) newInstance.invoke(null, rawData, offset, length);
+      return (IpV6Option) newInstance.invoke(null, rawData, offset, length);
     } catch (SecurityException e) {
       throw new IllegalStateException(e);
     } catch (NoSuchMethodException e) {
@@ -73,22 +75,22 @@ public final class PropertiesBasedIpV6RoutingDataFactory
       throw new IllegalStateException(e);
     } catch (InvocationTargetException e) {
       if (e.getTargetException() instanceof IllegalRawDataException) {
-        return IllegalIpV6RoutingData.newInstance(rawData, offset, length);
+        return IllegalIpV6Option.newInstance(rawData, offset, length);
       }
       throw new IllegalArgumentException(e);
     }
   }
 
   @Override
-  public Class<? extends IpV6RoutingData> getTargetClass(IpV6RoutingType type) {
-    if (type == null) {
-      throw new NullPointerException("type must not be null");
+  public Class<? extends IpV6Option> getTargetClass(IpV6OptionType number) {
+    if (number == null) {
+      throw new NullPointerException(" number: " + number);
     }
-    return PacketFactoryPropertiesLoader.getInstance().getIpV6RoutingDataClass(type);
+    return PacketFactoryPropertiesLoader.getInstance().getIpV6OptionClass(number);
   }
 
   @Override
-  public Class<? extends IpV6RoutingData> getTargetClass() {
-    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV6RoutingDataClass();
+  public Class<? extends IpV6Option> getTargetClass() {
+    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV6OptionClass();
   }
 }
