+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedIpV6RoutingDataFactory.java
@@ -5,39 +5,41 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import org.pcap4j.packet.IllegalIpV4Option;
+
+import org.pcap4j.packet.IllegalIpV6RoutingData;
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.IpV4Packet.IpV4Option;
-import org.pcap4j.packet.namednumber.IpV4OptionType;
+import org.pcap4j.packet.IpV6ExtRoutingPacket.IpV6RoutingData;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.namednumber.IpV6RoutingType;
 
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.14
  */
-public final class PropertiesBasedIpV4OptionFactory
-    implements PacketFactory<IpV4Option, IpV4OptionType> {
+public final class PropertiesBasedIpV6RoutingDataFactory
+    implements PacketFactory<IpV6RoutingData, IpV6RoutingType> {
 
-  private static final PropertiesBasedIpV4OptionFactory INSTANCE =
-      new PropertiesBasedIpV4OptionFactory();
+  private static final PropertiesBasedIpV6RoutingDataFactory INSTANCE =
+      new PropertiesBasedIpV6RoutingDataFactory();
 
-  private PropertiesBasedIpV4OptionFactory() {}
+  private PropertiesBasedIpV6RoutingDataFactory() {}
 
-  /** @return the singleton instance of PropertiesBasedIpV4OptionFactory. */
-  public static PropertiesBasedIpV4OptionFactory getInstance() {
+  /** @return the singleton instance of PropertiesBasedIpV6RoutingDataFactory. */
+  public static PropertiesBasedIpV6RoutingDataFactory getInstance() {
     return INSTANCE;
   }
 
   @Override
-  public IpV4Option newInstance(byte[] rawData, int offset, int length, IpV4OptionType number) {
-    return newInstance(rawData, offset, length, getTargetClass(number));
+  public IpV6RoutingData newInstance(byte[] rawData, int offset, int length, IpV6RoutingType type) {
+    return newInstance(rawData, offset, length, getTargetClass(type));
   }
 
   @Override
-  public IpV4Option newInstance(byte[] rawData, int offset, int length) {
+  public IpV6RoutingData newInstance(byte[] rawData, int offset, int length) {
     return newInstance(rawData, offset, length, getTargetClass());
   }
 
@@ -46,14 +48,14 @@ public final class PropertiesBasedIpV4OptionFactory
    * @param offset offset
    * @param length length
    * @param dataClass dataClass
-   * @return a new IpV4Option object.
+   * @return a new IpV6RoutingData object.
    * @throws IllegalStateException if an access to the newInstance method of the dataClass fails.
    * @throws IllegalArgumentException if an exception other than {@link IllegalRawDataException} is
    *     thrown by newInstance method of the dataClass.
    * @throws NullPointerException if any of arguments are null.
    */
-  public IpV4Option newInstance(
-      byte[] rawData, int offset, int length, Class<? extends IpV4Option> dataClass) {
+  public IpV6RoutingData newInstance(
+      byte[] rawData, int offset, int length, Class<? extends IpV6RoutingData> dataClass) {
     if (rawData == null || dataClass == null) {
       StringBuilder sb = new StringBuilder(50);
       sb.append("rawData: ").append(rawData).append(" dataClass: ").append(dataClass);
@@ -62,7 +64,7 @@ public final class PropertiesBasedIpV4OptionFactory
 
     try {
       Method newInstance = dataClass.getMethod("newInstance", byte[].class, int.class, int.class);
-      return (IpV4Option) newInstance.invoke(null, rawData, offset, length);
+      return (IpV6RoutingData) newInstance.invoke(null, rawData, offset, length);
     } catch (SecurityException e) {
       throw new IllegalStateException(e);
     } catch (NoSuchMethodException e) {
@@ -73,22 +75,22 @@ public final class PropertiesBasedIpV4OptionFactory
       throw new IllegalStateException(e);
     } catch (InvocationTargetException e) {
       if (e.getTargetException() instanceof IllegalRawDataException) {
-        return IllegalIpV4Option.newInstance(rawData, offset, length);
+        return IllegalIpV6RoutingData.newInstance(rawData, offset, length);
       }
       throw new IllegalArgumentException(e);
     }
   }
 
   @Override
-  public Class<? extends IpV4Option> getTargetClass(IpV4OptionType number) {
-    if (number == null) {
-      throw new NullPointerException("number: " + number);
+  public Class<? extends IpV6RoutingData> getTargetClass(IpV6RoutingType type) {
+    if (type == null) {
+      throw new NullPointerException("type must not be null");
     }
-    return PacketFactoryPropertiesLoader.getInstance().getIpV4OptionClass(number);
+    return PacketFactoryPropertiesLoader.getInstance().getIpV6RoutingDataClass(type);
   }
 
   @Override
-  public Class<? extends IpV4Option> getTargetClass() {
-    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV4OptionClass();
+  public Class<? extends IpV6RoutingData> getTargetClass() {
+    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV6RoutingDataClass();
   }
 }
