+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedRadiotapDataFactory.java
@@ -5,39 +5,42 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import org.pcap4j.packet.DnsResourceRecord.DnsRData;
-import org.pcap4j.packet.IllegalDnsRData;
+
+import org.pcap4j.packet.IllegalRadiotapData;
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.namednumber.DnsResourceRecordType;
+import org.pcap4j.packet.RadiotapPacket.RadiotapData;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.namednumber.RadiotapPresentBitNumber;
 
 /**
  * @author Kaito Yamada
- * @since pcap4j 1.7.1
+ * @since pcap4j 1.6.5
  */
-public final class PropertiesBasedDnsRDataFactory
-    implements PacketFactory<DnsRData, DnsResourceRecordType> {
+public final class PropertiesBasedRadiotapDataFactory
+    implements PacketFactory<RadiotapData, RadiotapPresentBitNumber> {
 
-  private static final PropertiesBasedDnsRDataFactory INSTANCE =
-      new PropertiesBasedDnsRDataFactory();
+  private static final PropertiesBasedRadiotapDataFactory INSTANCE =
+      new PropertiesBasedRadiotapDataFactory();
 
-  private PropertiesBasedDnsRDataFactory() {}
+  private PropertiesBasedRadiotapDataFactory() {}
 
-  /** @return the singleton instance of PropertiesBasedDnsRDataFactory. */
-  public static PropertiesBasedDnsRDataFactory getInstance() {
+  /** @return the singleton instance of PropertiesBasedRadiotapDataFactory. */
+  public static PropertiesBasedRadiotapDataFactory getInstance() {
     return INSTANCE;
   }
 
   @Override
-  public DnsRData newInstance(byte[] rawData, int offset, int length, DnsResourceRecordType num) {
+  public RadiotapData newInstance(
+      byte[] rawData, int offset, int length, RadiotapPresentBitNumber num) {
     return newInstance(rawData, offset, length, getTargetClass(num));
   }
 
   @Override
-  public DnsRData newInstance(byte[] rawData, int offset, int length) {
+  public RadiotapData newInstance(byte[] rawData, int offset, int length) {
     return newInstance(rawData, offset, length, getTargetClass());
   }
 
@@ -46,14 +49,14 @@ public final class PropertiesBasedDnsRDataFactory
    * @param offset offset
    * @param length length
    * @param dataClass dataClass
-   * @return a new DnsRDataField object.
+   * @return a new RadiotapDataField object.
    * @throws IllegalStateException if an access to the newInstance method of the dataClass fails.
    * @throws IllegalArgumentException if an exception other than {@link IllegalRawDataException} is
    *     thrown by newInstance method of the dataClass.
    * @throws NullPointerException if any of arguments are null.
    */
-  public DnsRData newInstance(
-      byte[] rawData, int offset, int length, Class<? extends DnsRData> dataClass) {
+  public RadiotapData newInstance(
+      byte[] rawData, int offset, int length, Class<? extends RadiotapData> dataClass) {
     if (rawData == null || dataClass == null) {
       StringBuilder sb = new StringBuilder(50);
       sb.append("rawData: ").append(rawData).append(" dataClass: ").append(dataClass);
@@ -62,7 +65,7 @@ public final class PropertiesBasedDnsRDataFactory
 
     try {
       Method newInstance = dataClass.getMethod("newInstance", byte[].class, int.class, int.class);
-      return (DnsRData) newInstance.invoke(null, rawData, offset, length);
+      return (RadiotapData) newInstance.invoke(null, rawData, offset, length);
     } catch (SecurityException e) {
       throw new IllegalStateException(e);
     } catch (NoSuchMethodException e) {
@@ -73,22 +76,22 @@ public final class PropertiesBasedDnsRDataFactory
       throw new IllegalStateException(e);
     } catch (InvocationTargetException e) {
       if (e.getTargetException() instanceof IllegalRawDataException) {
-        return IllegalDnsRData.newInstance(rawData, offset, length);
+        return IllegalRadiotapData.newInstance(rawData, offset, length);
       }
       throw new IllegalArgumentException(e);
     }
   }
 
   @Override
-  public Class<? extends DnsRData> getTargetClass(DnsResourceRecordType num) {
+  public Class<? extends RadiotapData> getTargetClass(RadiotapPresentBitNumber num) {
     if (num == null) {
       throw new NullPointerException("num must not be null");
     }
-    return PacketFactoryPropertiesLoader.getInstance().getDnsRDataClass(num);
+    return PacketFactoryPropertiesLoader.getInstance().getRadiotapDataFieldClass(num);
   }
 
   @Override
-  public Class<? extends DnsRData> getTargetClass() {
-    return PacketFactoryPropertiesLoader.getInstance().getUnknownDnsRDataClass();
+  public Class<? extends RadiotapData> getTargetClass() {
+    return PacketFactoryPropertiesLoader.getInstance().getUnknownRadiotapDataFieldClass();
   }
 }
