+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/propertiesbased/PropertiesBasedSctpChunkFactory.java
@@ -1,43 +1,45 @@
 /*_##########################################################################
   _##
-  _##  Copyright (C) 2012-2014  Pcap4J.org
+  _##  Copyright (C) 2016  Pcap4J.org
   _##
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.propertiesbased;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import org.pcap4j.packet.IllegalIpV6Option;
+
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.IpV6ExtOptionsPacket.IpV6Option;
-import org.pcap4j.packet.namednumber.IpV6OptionType;
+import org.pcap4j.packet.IllegalSctpChunk;
+import org.pcap4j.packet.SctpPacket.SctpChunk;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.namednumber.SctpChunkType;
 
 /**
  * @author Kaito Yamada
- * @since pcap4j 0.9.14
+ * @since pcap4j 1.6.6
  */
-public final class PropertiesBasedIpV6OptionFactory
-    implements PacketFactory<IpV6Option, IpV6OptionType> {
+public final class PropertiesBasedSctpChunkFactory
+    implements PacketFactory<SctpChunk, SctpChunkType> {
 
-  private static final PropertiesBasedIpV6OptionFactory INSTANCE =
-      new PropertiesBasedIpV6OptionFactory();
+  private static final PropertiesBasedSctpChunkFactory INSTANCE =
+      new PropertiesBasedSctpChunkFactory();
 
-  private PropertiesBasedIpV6OptionFactory() {}
+  private PropertiesBasedSctpChunkFactory() {}
 
-  /** @return the singleton instance of PropertiesBasedIpV6OptionFactory. */
-  public static PropertiesBasedIpV6OptionFactory getInstance() {
+  /** @return the singleton instance of PropertiesBasedSctpChunkFactory. */
+  public static PropertiesBasedSctpChunkFactory getInstance() {
     return INSTANCE;
   }
 
   @Override
-  public IpV6Option newInstance(byte[] rawData, int offset, int length, IpV6OptionType number) {
+  public SctpChunk newInstance(byte[] rawData, int offset, int length, SctpChunkType number) {
     return newInstance(rawData, offset, length, getTargetClass(number));
   }
 
   @Override
-  public IpV6Option newInstance(byte[] rawData, int offset, int length) {
+  public SctpChunk newInstance(byte[] rawData, int offset, int length) {
     return newInstance(rawData, offset, length, getTargetClass());
   }
 
@@ -46,14 +48,14 @@ public final class PropertiesBasedIpV6OptionFactory
    * @param offset offset
    * @param length length
    * @param dataClass dataClass
-   * @return a new IpV6Option object.
+   * @return a new SctpChunk object.
    * @throws IllegalStateException if an access to the newInstance method of the dataClass fails.
    * @throws IllegalArgumentException if an exception other than {@link IllegalRawDataException} is
    *     thrown by newInstance method of the dataClass.
    * @throws NullPointerException if any of arguments are null.
    */
-  public IpV6Option newInstance(
-      byte[] rawData, int offset, int length, Class<? extends IpV6Option> dataClass) {
+  public SctpChunk newInstance(
+      byte[] rawData, int offset, int length, Class<? extends SctpChunk> dataClass) {
     if (rawData == null || dataClass == null) {
       StringBuilder sb = new StringBuilder(50);
       sb.append("rawData: ").append(rawData).append(" dataClass: ").append(dataClass);
@@ -62,7 +64,7 @@ public final class PropertiesBasedIpV6OptionFactory
 
     try {
       Method newInstance = dataClass.getMethod("newInstance", byte[].class, int.class, int.class);
-      return (IpV6Option) newInstance.invoke(null, rawData, offset, length);
+      return (SctpChunk) newInstance.invoke(null, rawData, offset, length);
     } catch (SecurityException e) {
       throw new IllegalStateException(e);
     } catch (NoSuchMethodException e) {
@@ -73,22 +75,22 @@ public final class PropertiesBasedIpV6OptionFactory
       throw new IllegalStateException(e);
     } catch (InvocationTargetException e) {
       if (e.getTargetException() instanceof IllegalRawDataException) {
-        return IllegalIpV6Option.newInstance(rawData, offset, length);
+        return IllegalSctpChunk.newInstance(rawData, offset, length);
       }
       throw new IllegalArgumentException(e);
     }
   }
 
   @Override
-  public Class<? extends IpV6Option> getTargetClass(IpV6OptionType number) {
+  public Class<? extends SctpChunk> getTargetClass(SctpChunkType number) {
     if (number == null) {
-      throw new NullPointerException(" number: " + number);
+      throw new NullPointerException("number: " + number);
     }
-    return PacketFactoryPropertiesLoader.getInstance().getIpV6OptionClass(number);
+    return PacketFactoryPropertiesLoader.getInstance().getSctpChunkClass(number);
   }
 
   @Override
-  public Class<? extends IpV6Option> getTargetClass() {
-    return PacketFactoryPropertiesLoader.getInstance().getUnknownIpV6OptionClass();
+  public Class<? extends SctpChunk> getTargetClass() {
+    return PacketFactoryPropertiesLoader.getInstance().getUnknownSctpChunkClass();
   }
 }
