+++ pom.xml
@@ -147,8 +147,8 @@
           <version>3.1</version>
           <configuration>
             <encoding>UTF-8</encoding>
-            <source>1.6</source>
-            <target>1.6</target>
+            <source>1.7</source>
+            <target>1.7</target>
           </configuration>
         </plugin>
         <plugin>
